for all except for Merger4to1 series

+++ Merger2to1Z/Merger2to1Z.cpp
@@ -284,7 +284,7 @@ int Merger2to1Z::set_data(int data_byte_size)
     ///set OutPort buffer length
     m_out_data.data.length((unsigned int)data_byte_size + HEADER_BYTE_SIZE + FOOTER_BYTE_SIZE);
     memcpy(&(m_out_data.data[0]), &header[0], HEADER_BYTE_SIZE);
-    memcpy(&(m_out_data.data[HEADER_BYTE_SIZE]), &m_data1[0], data_byte_size);
+    memcpy(&(m_out_data.data[HEADER_BYTE_SIZE]), &m_data1[0], (size_t)data_byte_size);
     memcpy(&(m_out_data.data[HEADER_BYTE_SIZE + (unsigned int)data_byte_size]), &footer[0],
            FOOTER_BYTE_SIZE);
